+++ lib/open_meteo/types.rb
@@ -1,5 +1,8 @@
 require "dry-types"
 require "dry-struct"
+require "dry-validation"
+
+require_relative "application_contract"
 
 module OpenMeteo
   ##