+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Use [`dry-validation`](https://github.com/dry-rb/dry-validation) to validate data classes
 - Use [`dry-types`](https://github.com/dry-rb/dry-types) and [`dry-struct`](https://github.com/dry-rb/dry-struct) for data classes
 
 ## [0.0.2] - 2023-11-03